refactor(260427-dgh): extract snapshotBreadcrumbs to sibling file for HMR
CI lint failed on react-refresh/only-export-components — same precedent
as 260421-vxb. The component file must be a single-export module.

- Moved snapshotBreadcrumbs() to src/components/debug/snapshotBreadcrumbs.ts
- DebugAuthOverlay.tsx now imports from the sibling
- Test file updated to import from the new path

No behavior change. 389/389 tests pass; eslint and tsc clean.
